Move offline tests to a dedicated file.
