+++ tests/testsuite/build.rs
@@ -957,261 +957,6 @@ Did you mean `a`?",
 }
 
 #[test]
-fn cargo_compile_path_with_offline() {
-    let p = project()
-        .file(
-            "Cargo.toml",
-            r#"
-            [package]
-            name = "foo"
-            version = "0.0.1"
-            authors = []
-
-            [dependencies.bar]
-            path = "bar"
-        "#,
-        )
-        .file("src/lib.rs", "")
-        .file("bar/Cargo.toml", &basic_manifest("bar", "0.0.1"))
-        .file("bar/src/lib.rs", "")
-        .build();
-
-    p.cargo("build -Zoffline")
-        .masquerade_as_nightly_cargo()
-        .run();
-}
-
-#[test]
-fn cargo_compile_with_downloaded_dependency_with_offline() {
-    Package::new("present_dep", "1.2.3")
-        .file("Cargo.toml", &basic_manifest("present_dep", "1.2.3"))
-        .file("src/lib.rs", "")
-        .publish();
-
-    {
-        // make package downloaded
-        let p = project()
-            .file(
-                "Cargo.toml",
-                r#"
-            [project]
-            name = "foo"
-            version = "0.1.0"
-
-            [dependencies]
-            present_dep = "1.2.3"
-        "#,
-            )
-            .file("src/lib.rs", "")
-            .build();
-        p.cargo("build").run();
-    }
-
-    let p2 = project()
-        .at("bar")
-        .file(
-            "Cargo.toml",
-            r#"
-            [project]
-            name = "bar"
-            version = "0.1.0"
-
-            [dependencies]
-            present_dep = "1.2.3"
-        "#,
-        )
-        .file("src/lib.rs", "")
-        .build();
-
-    p2.cargo("build -Zoffline")
-        .masquerade_as_nightly_cargo()
-        .with_stderr(
-            "\
-[COMPILING] present_dep v1.2.3
-[COMPILING] bar v0.1.0 ([..])
-[FINISHED] dev [unoptimized + debuginfo] target(s) in [..]",
-        )
-        .run();
-}
-
-#[test]
-fn cargo_compile_offline_not_try_update() {
-    let p = project()
-        .at("bar")
-        .file(
-            "Cargo.toml",
-            r#"
-            [project]
-            name = "bar"
-            version = "0.1.0"
-
-            [dependencies]
-            not_cached_dep = "1.2.5"
-        "#,
-        )
-        .file("src/lib.rs", "")
-        .build();
-
-    p.cargo("build -Zoffline")
-        .masquerade_as_nightly_cargo()
-        .with_status(101)
-        .with_stderr(
-            "\
-error: no matching package named `not_cached_dep` found
-location searched: registry `[..]`
-required by package `bar v0.1.0 ([..])`
-As a reminder, you're using offline mode (-Z offline) \
-which can sometimes cause surprising resolution failures, \
-if this error is too confusing you may wish to retry \
-without the offline flag.",
-        )
-        .run();
-}
-
-#[test]
-fn compile_offline_without_maxvers_cached() {
-    Package::new("present_dep", "1.2.1").publish();
-    Package::new("present_dep", "1.2.2").publish();
-
-    Package::new("present_dep", "1.2.3")
-        .file("Cargo.toml", &basic_manifest("present_dep", "1.2.3"))
-        .file(
-            "src/lib.rs",
-            r#"pub fn get_version()->&'static str {"1.2.3"}"#,
-        )
-        .publish();
-
-    Package::new("present_dep", "1.2.5")
-        .file("Cargo.toml", &basic_manifest("present_dep", "1.2.5"))
-        .file("src/lib.rs", r#"pub fn get_version(){"1.2.5"}"#)
-        .publish();
-
-    {
-        // make package cached
-        let p = project()
-            .file(
-                "Cargo.toml",
-                r#"
-            [project]
-            name = "foo"
-            version = "0.1.0"
-
-            [dependencies]
-            present_dep = "=1.2.3"
-        "#,
-            )
-            .file("src/lib.rs", "")
-            .build();
-        p.cargo("build").run();
-    }
-
-    let p2 = project()
-        .file(
-            "Cargo.toml",
-            r#"
-            [project]
-            name = "foo"
-            version = "0.1.0"
-
-            [dependencies]
-            present_dep = "1.2"
-        "#,
-        )
-        .file(
-            "src/main.rs",
-            "\
-extern crate present_dep;
-fn main(){
-    println!(\"{}\", present_dep::get_version());
-}",
-        )
-        .build();
-
-    p2.cargo("run -Zoffline")
-        .masquerade_as_nightly_cargo()
-        .with_stderr(
-            "\
-[COMPILING] present_dep v1.2.3
-[COMPILING] foo v0.1.0 ([CWD])
-[FINISHED] dev [unoptimized + debuginfo] target(s) in [..]
-     Running `[..]`",
-        )
-        .with_stdout("1.2.3")
-        .run();
-}
-
-#[test]
-fn offline_unused_target_dep() {
-    // -Z offline with a target dependency that is not used and not downloaded.
-    Package::new("unused_dep", "1.0.0").publish();
-    Package::new("used_dep", "1.0.0").publish();
-    let p = project()
-        .file(
-            "Cargo.toml",
-            r#"
-            [project]
-            name = "foo"
-            version = "0.1.0"
-            [dependencies]
-            used_dep = "1.0"
-            [target.'cfg(unused)'.dependencies]
-            unused_dep = "1.0"
-            "#,
-        )
-        .file("src/lib.rs", "")
-        .build();
-    // Do a build that downloads only what is necessary.
-    p.cargo("build")
-        .with_stderr_contains("[DOWNLOADED] used_dep [..]")
-        .with_stderr_does_not_contain("[DOWNLOADED] unused_dep [..]")
-        .run();
-    p.cargo("clean").run();
-    // Build offline, make sure it works.
-    p.cargo("build -Z offline")
-        .masquerade_as_nightly_cargo()
-        .run();
-}
-
-#[test]
-fn offline_missing_optional() {
-    Package::new("opt_dep", "1.0.0").publish();
-    let p = project()
-        .file(
-            "Cargo.toml",
-            r#"
-            [project]
-            name = "foo"
-            version = "0.1.0"
-            [dependencies]
-            opt_dep = { version = "1.0", optional = true }
-            "#,
-        )
-        .file("src/lib.rs", "")
-        .build();
-    // Do a build that downloads only what is necessary.
-    p.cargo("build")
-        .with_stderr_does_not_contain("[DOWNLOADED] opt_dep [..]")
-        .run();
-    p.cargo("clean").run();
-    // Build offline, make sure it works.
-    p.cargo("build -Z offline")
-        .masquerade_as_nightly_cargo()
-        .run();
-    p.cargo("build -Z offline --features=opt_dep")
-        .masquerade_as_nightly_cargo()
-        .with_stderr(
-            "\
-[ERROR] failed to download `opt_dep v1.0.0`
-
-Caused by:
-  can't make HTTP request in the offline mode
-",
-        )
-        .with_status(101)
-        .run();
-}
-
-#[test]
 fn incompatible_dependencies() {
     Package::new("bad", "0.1.0").publish();
     Package::new("bad", "1.0.0").publish();
@@ -1306,58 +1051,6 @@ all possible versions conflict with previously selected packages.
     ... which is depended on by `foo v0.0.1 ([..])`
 
 failed to select a version for `bad` which could resolve this conflict",
-        )
-        .run();
-}
-
-#[test]
-fn compile_offline_while_transitive_dep_not_cached() {
-    let baz = Package::new("baz", "1.0.0");
-    let baz_path = baz.archive_dst();
-    baz.publish();
-
-    let mut content = Vec::new();
-
-    let mut file = File::open(baz_path.clone()).ok().unwrap();
-    let _ok = file.read_to_end(&mut content).ok().unwrap();
-    drop(file);
-    drop(File::create(baz_path.clone()).ok().unwrap());
-
-    Package::new("bar", "0.1.0").dep("baz", "1.0.0").publish();
-
-    let p = project()
-        .file(
-            "Cargo.toml",
-            r#"
-            [project]
-            name = "foo"
-            version = "0.0.1"
-
-            [dependencies]
-            bar = "0.1.0"
-        "#,
-        )
-        .file("src/main.rs", "fn main(){}")
-        .build();
-
-    // simulate download bar, but fail to download baz
-    p.cargo("build")
-        .with_status(101)
-        .with_stderr_contains("[..]failed to verify the checksum of `baz[..]")
-        .run();
-
-    drop(File::create(baz_path).ok().unwrap().write_all(&content));
-
-    p.cargo("build -Zoffline")
-        .masquerade_as_nightly_cargo()
-        .with_status(101)
-        .with_stderr(
-            "\
-[ERROR] failed to download `baz v1.0.0`
-
-Caused by:
-  can't make HTTP request in the offline mode
-",
         )
         .run();
 }
