+++ tests/testsuite/git.rs
@@ -84,175 +84,6 @@ fn cargo_compile_simple_git_dep() {
 }
 
 #[test]
-fn cargo_compile_forbird_git_httpsrepo_offline() {
-    let p = project()
-        .file(
-            "Cargo.toml",
-            r#"
-
-            [project]
-            name = "foo"
-            version = "0.5.0"
-            authors = ["chabapok@example.com"]
-
-            [dependencies.dep1]
-            git = 'https://github.com/some_user/dep1.git'
-        "#,
-        )
-        .file("src/main.rs", "")
-        .build();
-
-    p.cargo("build -Zoffline").masquerade_as_nightly_cargo().with_status(101).
-                    with_stderr("\
-error: failed to load source for a dependency on `dep1`
-
-Caused by:
-  Unable to update https://github.com/some_user/dep1.git
-
-Caused by:
-  can't checkout from 'https://github.com/some_user/dep1.git': you are in the offline mode (-Z offline)").run();
-}
-
-#[test]
-fn cargo_compile_offline_with_cached_git_dep() {
-    let git_project = git::new("dep1", |project| {
-        project
-            .file("Cargo.toml", &basic_lib_manifest("dep1"))
-            .file(
-                "src/lib.rs",
-                r#"
-                pub static COOL_STR:&str = "cached git repo rev1";
-            "#,
-            )
-    })
-    .unwrap();
-
-    let repo = git2::Repository::open(&git_project.root()).unwrap();
-    let rev1 = repo.revparse_single("HEAD").unwrap().id();
-
-    // Commit the changes and make sure we trigger a recompile
-    File::create(&git_project.root().join("src/lib.rs"))
-        .unwrap()
-        .write_all(br#"pub static COOL_STR:&str = "cached git repo rev2";"#)
-        .unwrap();
-    git::add(&repo);
-    let rev2 = git::commit(&repo);
-
-    {
-        // cache to registry rev1 and rev2
-        let prj = project()
-            .at("cache_git_dep")
-            .file(
-                "Cargo.toml",
-                &format!(
-                    r#"
-            [project]
-            name = "cache_git_dep"
-            version = "0.5.0"
-
-            [dependencies.dep1]
-            git = '{}'
-            rev = "{}"
-            "#,
-                    git_project.url(),
-                    rev1
-                ),
-            )
-            .file("src/main.rs", "fn main(){}")
-            .build();
-        prj.cargo("build").run();
-
-        File::create(&prj.root().join("Cargo.toml"))
-            .unwrap()
-            .write_all(
-                format!(
-                    r#"
-            [project]
-            name = "cache_git_dep"
-            version = "0.5.0"
-
-            [dependencies.dep1]
-            git = '{}'
-            rev = "{}"
-            "#,
-                    git_project.url(),
-                    rev2
-                )
-                .as_bytes(),
-            )
-            .unwrap();
-        prj.cargo("build").run();
-    }
-
-    let p = project()
-        .file(
-            "Cargo.toml",
-            &format!(
-                r#"
-            [project]
-            name = "foo"
-            version = "0.5.0"
-
-            [dependencies.dep1]
-            git = '{}'
-        "#,
-                git_project.url()
-            ),
-        )
-        .file(
-            "src/main.rs",
-            &main_file(r#""hello from {}", dep1::COOL_STR"#, &["dep1"]),
-        )
-        .build();
-
-    let git_root = git_project.root();
-
-    p.cargo("build -Zoffline")
-        .masquerade_as_nightly_cargo()
-        .with_stderr(format!(
-            "\
-[COMPILING] dep1 v0.5.0 ({}#[..])
-[COMPILING] foo v0.5.0 ([CWD])
-[FINISHED] dev [unoptimized + debuginfo] target(s) in [..]",
-            path2url(git_root),
-        ))
-        .run();
-
-    assert!(p.bin("foo").is_file());
-
-    p.process(&p.bin("foo"))
-        .with_stdout("hello from cached git repo rev2\n")
-        .run();
-
-    File::create(&p.root().join("Cargo.toml"))
-        .unwrap()
-        .write_all(
-            format!(
-                r#"
-        [project]
-        name = "foo"
-        version = "0.5.0"
-
-        [dependencies.dep1]
-        git = '{}'
-        rev = "{}"
-    "#,
-                git_project.url(),
-                rev1
-            )
-            .as_bytes(),
-        )
-        .unwrap();
-
-    p.cargo("build -Zoffline")
-        .masquerade_as_nightly_cargo()
-        .run();
-    p.process(&p.bin("foo"))
-        .with_stdout("hello from cached git repo rev1\n")
-        .run();
-}
-
-#[test]
 fn cargo_compile_git_dep_branch() {
     let project = project();
     let git_project = git::new("dep1", |project| {
