+++ tests/testsuite/registry.rs
@@ -820,30 +820,6 @@ fn update_lockfile() {
 }
 
 #[test]
-fn update_offline() {
-    let p = project()
-        .file(
-            "Cargo.toml",
-            r#"
-            [project]
-            name = "foo"
-            version = "0.0.1"
-            authors = []
-
-            [dependencies]
-            bar = "*"
-        "#,
-        )
-        .file("src/main.rs", "fn main() {}")
-        .build();
-    p.cargo("update -Zoffline")
-        .masquerade_as_nightly_cargo()
-        .with_status(101)
-        .with_stderr("error: you can't update in the offline mode[..]")
-        .run();
-}
-
-#[test]
 fn dev_dependency_not_used() {
     let p = project()
         .file(
